UPSTREAM: arm64: lse: Pass -fomit-frame-pointer to out-of-line ll/sc … …atomics
In cases where x30 is used as a temporary in the out-of-line ll/sc atomics
(e.g. atomic_fetch_add), the compiler tends to put out a full stackframe,
which included pointing the x29 at the new frame.

Since these things aren't traceable anyway, we can pass -fomit-frame-pointer
to reduce the work when spilling. Since this is incompatible with -pg, we
also remove that from the CFLAGS for this file.

Bug: 122733725
Change-Id: I23939e1524040e8ab6ed0845b34183609bab079d
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: AnuragRai <kerneldeveloper7232@gmail.com>